refactor(api): create base framework controller
